# server/docs/postman-openapi-full.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Zalo Clone — Chat Server API (full)",
    "version": "1.0.0",
...
  },
  "servers": [
    {
      "url": "https://unlimited-feedback-crossing-academics.trycloudflare.com",
      "description": "Qua Cloudflare tunnel (gọi từ máy khác)"
    },
    {
      "url": "http://localhost:4000",
      "description": "Local trực tiếp tại máy chủ"
    }
  ],
...
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "AGENT_API_KEY hoặc ADMIN_API_KEY trong `server/.env`. GET /api/media/* được miễn."
      }
    },
...
    "responses": {
...
      "Unauthorized": {
        "description": "Thiếu hoặc sai API key (header `X-Api-Key`).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Mô tả lỗi (tiếng Việt)."
          },
          "code": {
            "type": "string",
            "nullable": true,
            "description": "Mã lỗi gốc từ Zalo (nếu có)."
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Passthrough": {
        "type": "object",
        "additionalProperties": true,
        "description": "Object thô trả về từ Zalo (zca-js). Cấu trúc tuỳ thao tác."
      },
...
    }
  },
  "paths": {
...
    "/api/friends/requests/sent": {
      "get": {
        "tags": [
          "Bạn bè"
        ],
        "summary": "Lời mời kết bạn đã gửi",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/friends/recommendations": {
      "get": {
        "tags": [
          "Bạn bè"
        ],
        "summary": "Gợi ý kết bạn",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/friends/online": {
      "get": {
        "tags": [
          "Bạn bè"
        ],
        "summary": "Bạn bè đang online",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/friends/aliases": {
      "get": {
        "tags": [
          "Bạn bè"
        ],
        "summary": "Danh sách biệt danh",
        "parameters": [
          {
            "name": "count",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/friends/find": {
      "get": {
        "tags": [
          "Bạn bè"
        ],
        "summary": "Tìm người dùng theo SĐT",
        "parameters": [
          {
            "name": "phone",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "inbox"
      }
    },
    "/api/friends/requests": {
      "post": {
        "tags": [
          "Bạn bè"
        ],
        "summary": "Gửi lời mời kết bạn",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "msg": {
                    "type": "string"
                  }
                },
                "required": [
                  "userId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/friends/requests/{friendId}/accept": {
      "post": {
        "tags": [
          "Bạn bè"
        ],
        "summary": "Chấp nhận kết bạn",
        "parameters": [
          {
            "name": "friendId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/friends/requests/{friendId}/reject": {
      "post": {
        "tags": [
          "Bạn bè"
        ],
        "summary": "Từ chối kết bạn",
        "parameters": [
          {
            "name": "friendId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/friends/requests/{friendId}/undo": {
      "post": {
        "tags": [
          "Bạn bè"
        ],
        "summary": "Thu hồi lời mời đã gửi",
        "parameters": [
          {
            "name": "friendId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/friends/{friendId}/status": {
      "get": {
        "tags": [
          "Bạn bè"
        ],
        "summary": "Trạng thái lời mời kết bạn",
        "parameters": [
          {
            "name": "friendId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/friends/{friendId}/profile": {
      "get": {
        "tags": [
          "Bạn bè"
        ],
        "summary": "Hồ sơ người dùng",
        "parameters": [
          {
            "name": "friendId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "inbox"
      }
    },
    "/api/friends/{friendId}/related-groups": {
      "get": {
        "tags": [
          "Bạn bè"
        ],
        "summary": "Nhóm chung với người này",
        "parameters": [
          {
            "name": "friendId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/friends/{friendId}/block": {
      "post": {
        "tags": [
          "Bạn bè"
        ],
        "summary": "Chặn người dùng",
        "parameters": [
          {
            "name": "friendId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/friends/{friendId}/unblock": {
      "post": {
        "tags": [
          "Bạn bè"
        ],
        "summary": "Bỏ chặn người dùng",
        "parameters": [
          {
            "name": "friendId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/friends/{friendId}/alias": {
      "put": {
        "tags": [
          "Bạn bè"
        ],
        "summary": "Đặt biệt danh",
        "parameters": [
          {
            "name": "friendId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "alias": {
                    "type": "string"
                  }
                },
                "required": [
                  "alias"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      },
      "delete": {
        "tags": [
          "Bạn bè"
        ],
        "summary": "Xoá biệt danh",
        "parameters": [
          {
            "name": "friendId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
    "/api/friends/{friendId}": {
      "delete": {
        "tags": [
          "Bạn bè"
        ],
        "summary": "Xoá bạn",
        "parameters": [
          {
            "name": "friendId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Passthrough"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "x-audience": "internal"
      }
    },
...
  }
}